Rename CardsAPI directory
