+++ src/humans/Dealer.py
@@ -1,4 +1,4 @@
-from src.CardsAPI.hand import Hand
+from src.cards.hand import Hand
 from src.common.constants import Decision
