+++ src/views/tests/test_card_area_organizer.py
@@ -1,4 +1,4 @@
-from src.CardsAPI.card import Card
+from src.cards.card import Card
 from src.common.game_view_config import game_view_config
 from src.views.card_area_organizer import CardAreaOrganizer
